% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colnameIDs_to_Names}
\alias{colnameIDs_to_Names}
\title{colnameIDs_to_Names}
\usage{
colnameIDs_to_Names(df.update, df.LinkID_Name, ID, Name)
}
\arguments{
\item{df.update}{number of columns}

\item{df.LinkID_Name}{column names}

\item{ID}{column names}

\item{Name}{column names}
}
\description{
create an empty dataframe
}
\examples{
colnameIDs_to_Names(n.cols,col.names)
}
