% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_empty_df}
\alias{create_empty_df}
\title{create_empty_df}
\usage{
create_empty_df(n.cols, col.names)
}
\arguments{
\item{n.cols}{number of columns}

\item{col.names}{column names}
}
\description{
create an empty dataframe
}
\examples{
create_empty_df(n.cols,col.names)
}
