% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_IDstring.R
\name{create_IDstring}
\alias{create_IDstring}
\title{create_IDstring}
\usage{
create_IDstring(numeric_vector)
}
\arguments{
\item{numeric_vector}{ndexes to pull a subset.}
}
\description{
helper function for writing SQL querries. Convers a numeric vector into character string typically used in WHERE IN statement
}
\examples{
statement<-paste("SELECT * FROM Alloc WHERE ItemID IN ",create_IDstring(Items$ItemID),";",sep='')
}
