% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{db_clean}
\alias{db_clean}
\title{db_clean}
\usage{
db_clean(df)
}
\arguments{
\item{x}{string to trim}
}
\description{
df data frame to clean, will eliminate leading and trailing spaces and also prep the data for upload by taking care of apostrophe's
}
\examples{
db_clean(df)
}

