% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doUpdateFullQueryAnyAny_MS.R
\name{doUpdateFullQueryAnyAny_MS}
\alias{doUpdateFullQueryAnyAny_MS}
\title{doUpdateFullQueryAnyAny_MS}
\usage{
doUpdateFullQueryAnyAny_MS(con, tableName, dfFull, updateFields, batchSize)
}
\arguments{
\item{con}{RMySQL connection}

\item{tableName}{Table name to update}

\item{dfFull}{data frame to update}

\item{updateFields}{subset of fields to update or NULL for all fields}

\item{batchSize}{number of rows to process, works well up to 2000, generally good up to 5000}
}
\description{
SQL table update function
}
\examples{
statement<-paste("SELECT * FROM Alloc WHERE ItemID IN ",create_IDstring(Items$ItemID),";",sep='')
}
