% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DownloadProgramInventory_Edit.R
\name{DownloadProgramInventory_Edit}
\alias{DownloadProgramInventory_Edit}
\title{DownloadProgramInventory_Edit}
\usage{
DownloadProgramInventory_Edit(fname, BudgetID = NULL, ItemMeta1 = NULL,
  db_name, CostModelInfo, CostModelID, include.sheets = c("Program Inventory",
  "BudgetInfo", "ProgBudgetInfo"))
}
\arguments{
\item{fname}{If using for download template provide this}

\item{BudgetID}{Table name to update}

\item{ItemMeta1}{Subset by a user group}

\item{db_name}{}

\item{CostModelInfo}{The cost model info table}

\item{CostModelID}{Costmodel ID to use, typically the one from the PBB cost model}

\item{include.sheets}{Which sheets to include in the list of standard templates}
}
\description{
Some standard templates for ProgInventory and Costs
}
\examples{
ProgInfo<-DownloadProgramInventory_Edit(fname=NULL,BudgetID=NULL,ItemMeta1=NULL,CostModelInfo,CostModelID)
}

