% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_AllPrograms.R
\name{get_AllPrograms}
\alias{get_AllPrograms}
\title{get_AllPrograms}
\usage{
get_AllPrograms(BudgetID = NULL, ItemMeta1 = NULL, db_name, CostModelInfo,
  CostModelID)
}
\arguments{
\item{BudgetID}{If using for download template provide this}

\item{ItemMeta1}{}

\item{db_name}{data base to use}

\item{CostModelInfo}{The cost model info table}

\item{CostModelID}{Costmodel ID to use, typically the one from the PBB cost model}
}
\description{
Function to pull programs in onlinepbb.net/home
}
\examples{
ProgInfo<-get_AllPrograms(BudgetID=NULL, ItemMeta1=NULL,db_name,CostModelInfo,CostModelID)
}

