% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_BudgetProgIDs}
\alias{get_BudgetProgIDs}
\title{selectInput_filter}
\usage{
get_BudgetProgIDs(AppBudgetID, conexists = NULL,
  Div1Name = values.setup$CostModelInfo$Div1Name,
  Div2Name = values.setup$CostModelInfo$Div2Name)
}
\arguments{
\item{AppBudgetID}{dataframe to subset}

\item{conexists}{input filter}

\item{Div1Name}{column to filter on}

\item{Div2Name}{include all term, example "All Available"}
}
\description{
subset a dataframe from select input
}
\examples{
create_empty_df(n.cols,col.names)
}
