% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FinalScores.R
\name{get_FinalScores}
\alias{get_FinalScores}
\title{get_ProgramCosts}
\usage{
get_FinalScores(db_name, All.Programs = F)
}
\arguments{
\item{db_name}{Table name to update}

\item{BudgetID}{BudegtID to pull}
}
\description{
Get Program Inventory Costs and Revenue
}
\examples{
data<-get_ProgramCosts(BudgetID=1,db_name='RX_Training5')
}
