% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ProgramCosts.R
\name{get_ProgramCosts}
\alias{get_ProgramCosts}
\title{get_ProgramCosts}
\usage{
get_ProgramCosts(db_name, BudgetID, ItemIDs = NULL,
  UserGroups = "All Available")
}
\arguments{
\item{db_name}{Table name to update}

\item{BudgetID}{BudegtID to pull}
}
\description{
Get Program Inventory Costs and Revenue
}
\examples{
data<-get_ProgramCosts(BudgetID=1,db_name='RX_Training5')
}
