% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RadSlicer.R
\name{get_RadSlicer}
\alias{get_RadSlicer}
\title{get_RadSlicer}
\usage{
get_RadSlicer(db_name, BudgetIDs)
}
\arguments{
\item{db_name}{Table name to update}

\item{BudgetID}{BudegtID to pull}
}
\description{
Get Program Inventory Costs and Revenue
}
\examples{
data<-get_RadSlicer(BudgetID=1,db_name='RX_Training5')
}
