% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectInput_filter}
\alias{selectInput_filter}
\title{selectInput_filter}
\usage{
selectInput_filter(df, filter, col.filter, all.option, by.rows = NULL,
  includeBlanckDivs = NULL)
}
\arguments{
\item{df}{dataframe to subset}

\item{filter}{input filter}

\item{col.filter}{column to filter on}

\item{all.option}{include all term, example "All Available"}

\item{by.rows}{if NULL then normal filtering, if FALSE then filter by columns. USed in Edit USers filtering.}

\item{includeBlanckDivs}{not sure if this is used anymore}
}
\description{
subset a dataframe from select input
}
\examples{
create_empty_df(n.cols,col.names)
}
