% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_InvandCosts.R
\name{get_InvandCosts}
\alias{get_InvandCosts}
\title{get_InvandCosts}
\usage{
get_InvandCosts(BudgetID, db_name, CostModelName = "PBB", ItemMeta1 = NULL)
}
\arguments{
\item{BudgetID}{BudegtID to pull}

\item{db_name}{Table name to update}

\item{CostModelName}{Typically "PBB"}
}
\description{
Get Program Inventory Costs and Revenue
}
\examples{
data<-get_InvandCosts(BudgetID=1,db_name='RX_Training5')
}

